Functional version of the 24HourWeeklyProductivityLog
Copy and paste to the end the TEMPLATE page and fill in the first date at the top, then drag across to fill in the blanks.

Fill in each cell as 30 minute increments, I just round up to the full half-hour if whatever task I'm on takes at least 15 minutes. Use your better judgement though, this can be as loose or tight as you want it!

I made this sheet to get into the habbit of tracking my study times and stick to a healthy schedule as I balance school/ side learning and working out and self-care. Make sure you include that latter part! It's important to give your mind a break often to increase your yields.

As I finish out the year I plan on ironing out the good and the bad, then merging the new version for the next year.

Anyway this version should provide some graphical representation to your time blocks if you stick to naming the cells after the title in the color legend. You'll see my example on page '49'. By the way, I personally name my pages after the week number in the year, but I'm sure you can name them whatever you want!

Enjoy!!!